refactor: TabContents 이름수정
