+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as TabContent } from "./TabContent";
+export { default as TabContents } from "./TabContents";
 export { default as Layout } from "./Layout";
 export { default as Footer } from "./Footer";
 export { default as DataDummy } from "./DataDummy";
